backend f/requirements.txt
Flask
numpy
scikit-learn
xgboost
joblib
psycopg2-binary
python-dotenv
